feat(server): fold LoroSyncBroadcaster into CommitMonitor
Ephemera and drive presence share the commit-monitor mailbox. One
UnsubscribeAll on socket close walks resource, drive, Loro, and presence
maps. AppState and the WebSocket actor hold a single Addr. Wire and
fan-out behaviour are unchanged.
